# docs/swagger.yaml
basePath: /api
definitions:
...
  dtos.PatchStockProductRequest:
    properties:
      quantity:
        type: integer
      unit_cost:
        type: number
    type: object
...
  dtos.ProductsImageDto:
    properties:
      id:
        type: string
      image_url:
        type: string
      product_id:
        type: string
    type: object
  dtos.PutStockProductRequest:
    properties:
      quantity:
        type: integer
      unit_cost:
        type: number
    type: object
...
  dtos.StockProductInfo:
    properties:
      image:
        $ref: '#/definitions/dtos.ProductsImageDto'
      name:
        type: string
      product_id:
        type: string
      quantity:
        type: integer
      unit_cost:
        type: number
    type: object
...
host: marketplace-vendor-products-service.onrender.com
info:
  contact: {}
  description: API for managing products and stocks for vendors.
  title: Product Service API
  version: "1.0"
paths:
...
  /stocks/{stockId}/products/{productId}:
    delete:
      consumes:
      - application/json
      description: Deletes a specific product from a given stock for the vendor.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Stock ID (UUID)
        in: path
        name: stockId
        required: true
        type: string
      - description: Product ID (UUID)
        in: path
        name: productId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid vendorId/stockId/productId
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Stock product not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete product from stock by ID
      tags:
      - stocks
    patch:
      consumes:
      - application/json
      description: Partially updates the details of a specific product within a given
        stock for the vendor.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Stock ID (UUID)
        in: path
        name: stockId
        required: true
        type: string
      - description: Product ID (UUID)
        in: path
        name: productId
        required: true
        type: string
      - description: Fields to update in the stock product
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.PatchStockProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.StockProductInfo'
        "400":
          description: Invalid vendorId/stockId/productId/stock product data
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Stock product not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Partially update a product in a stock by IDs
      tags:
      - stocks
    put:
      consumes:
      - application/json
      description: Modifies a specific product within a given stock for the vendor.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Stock ID (UUID)
        in: path
        name: stockId
        required: true
        type: string
      - description: Product ID (UUID)
        in: path
        name: productId
        required: true
        type: string
      - description: Full product data for replacement
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.PutStockProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid vendorId/stockId/productId/stock product data
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Stock product not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Modify a product in a stock by IDs
      tags:
      - stocks
schemes:
- https
swagger: "2.0"